+++ Changes.md
@@ -1,6 +1,9 @@
 #  Changes
 
 ## v1.2
+- Fixed Window State Problem
+- Fixed the search field
+- Fixed the bug when closing the first tab, it will close the window
 - Created run script to make life easier :)
 - Fixed the bug where if you close the first tab it will crash